Fixed task 015, now it gives an int

+++ 015 Lattice paths/main.py
@@ -3,7 +3,7 @@
 import math
 
 def getNumPaths(n):
-	return math.factorial(2*n)/math.pow(math.factorial(n),2)
+	return int(math.factorial(2*n)/math.pow(math.factorial(n),2))
 	
 
 print(str(getNumPaths(20)))
